Splitting up CategoricalOutput & ContrastiveOutput
