+++ merlin/models/tf/__init__.py
@@ -114,7 +114,8 @@ from merlin.models.tf.models.retrieval import (
     YoutubeDNNRetrievalModel,
 )
 from merlin.models.tf.outputs.base import ModelOutput
-from merlin.models.tf.outputs.classification import BinaryOutput, CategoricalPrediction
+from merlin.models.tf.outputs.classification import BinaryOutput, CategoricalOutput
+from merlin.models.tf.outputs.contrastive import ContrastiveOutput
 from merlin.models.tf.outputs.regression import RegressionOutput
 from merlin.models.tf.outputs.sampling.base import Items, ItemSamplerV2
 from merlin.models.tf.outputs.sampling.in_batch import InBatchSamplerV2
@@ -192,7 +193,8 @@ __all__ = [
     "ModelOutput",
     "BinaryOutput",
     "RegressionOutput",
-    "CategoricalPrediction",
+    "CategoricalOutput",
+    "ContrastiveOutput",
     "PredictionTask",
     "BinaryClassificationTask",
     "MultiClassClassificationTask",
